Remove unnecessary files and update package name
