--- README.md
@@ -1,1 +1,0 @@
-# mars_rover
