+++ src/domain/obstacle.ts
