+++ src/domain/planet.interface.ts
@@ -1,0 +1,7 @@
+import { Point } from "./point";
+
+export interface PlanetInterface {
+    Normaliser(point: Point): Point;
+    EstLibre(position: Point): boolean;
+    isObstacle(point: Point): boolean;
+}
