+++ src/domain/point.ts
