--- src/obstacle.js
@@ -1,10 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.Obstacle = void 0;
-var Obstacle = /** @class */ (function () {
-    function Obstacle(position) {
-        this.position = position;
-    }
-    return Obstacle;
-}());
-exports.Obstacle = Obstacle;
